new logic for LogController and new Helper
