+++ app/Http/Middleware/TokenCheck.php
@@ -2,14 +2,15 @@
 
 namespace App\Http\Middleware;
 
-use App\Helpers\Init;
+use App\Helpers\ServiceManager;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
-use Carbon\Carbon;
 use Validator;
 
 class TokenCheck
 {
+    private const LOG_MESSAGE = 'TokenCheck::sendResult TOKEN ERROR FOR';
+
     /**
      * handle - обрабатывает токен для записи лога
      * 
@@ -43,19 +44,27 @@ class TokenCheck
             return response()->json($validate->errors(), 400);
         }
         $data = $request->all();
-        $parcedData = Init::returnParts($data);
 
-        $checkResult = match ($parcedData['service']) {
+        // Проверим что service имеет вид "service|type"
+        $parcedData = ServiceManager::returnParts($data);
+        if (!$parcedData['success']) {
+            Log::channel("tokens")->info(self::LOG_MESSAGE, $userData);
+            return response()->json(['success' => false, "message" => "Неверно указан сервис"]);
+        }
+
+        // Вызываем соответствующий метод проверки токена
+        $checkResult = match ($parcedData['data']['service']) {
             'WSPG' => $this->checkTokenForWsPg($data),
             'ADS' => false,
             default => false,
         };
 
         if (!$checkResult) {
-            Log::channel("tokens")->info('\TokenCheck::sendResult TOKEN ERROR FOR', $userData);
-            return response()->json(['success' => false, "message" => "Неверно указан сервис"]);
+            Log::channel("tokens")->info(self::LOG_MESSAGE, $userData);
+            return response()->json(['success' => false, "message" => "Неверный токен"]);
         }
 
+        // Все ок, го некст
         Log::channel("tokens")->info('TokenCheck::sendResult USER IS AUTHORIZED', $userData);
         return $next($request);
     }
